fix(worker): wire platform_store in DurableWorker act_activity path
The DurableWorker (used in full/production mode) calls act_activity()
from activities.rs which constructed ActAtom without platform_store.
This caused "Platform management not available in this context" errors
for manage_harnesses and other platform management tools.

The unified_worker.rs path (used in dev mode) already wired
platform_store correctly, which is why the bug only manifested in
full mode with separate worker processes.

- Add GrpcPlatformStore to ActAtom in activities.rs
- Extract shared MockPlatformStore to platform_store.rs for reuse
- Add ActAtom-level integration tests that verify platform_store wiring
- Refactor platform_management.rs tests to use shared mock
