+++ crates/core/src/capabilities/platform_management.rs
@@ -1001,216 +1001,10 @@ impl Tool for SessionInteractTool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::AgentCapabilityConfig;
-    use crate::agent::{Agent, AgentStatus};
-    use crate::error::Result;
-    use crate::harness::{Harness, HarnessStatus};
-    use crate::platform_store::{PlatformMessage, PlatformStore};
-    use crate::session::{Session, SessionStatus};
+    use crate::platform_store::PlatformStore;
+    use crate::platform_store::tests::MockPlatformStore;
     use crate::typed_id::{AgentId, HarnessId, SessionId};
-    use async_trait::async_trait;
-    use chrono::Utc;
     use std::sync::Arc;
-
-    fn build_harness(name: &str) -> Harness {
-        Harness {
-            id: HarnessId::new(),
-            name: name.to_string(),
-            description: Some("test harness".to_string()),
-            system_prompt: "You are helpful.".to_string(),
-            default_model_id: None,
-            tags: vec![],
-            capabilities: vec![AgentCapabilityConfig::new("session")],
-            status: HarnessStatus::Active,
-            created_at: Utc::now(),
-            updated_at: Utc::now(),
-        }
-    }
-
-    fn build_agent(name: &str) -> Agent {
-        Agent {
-            public_id: AgentId::new(),
-            internal_id: uuid::Uuid::now_v7(),
-            name: name.to_string(),
-            description: Some("test agent".to_string()),
-            system_prompt: "You are helpful.".to_string(),
-            default_model_id: None,
-            tags: vec![],
-            capabilities: vec![],
-            tools: vec![],
-            status: AgentStatus::Active,
-            created_at: Utc::now(),
-            updated_at: Utc::now(),
-            usage: None,
-        }
-    }
-
-    fn build_session(title: &str) -> Session {
-        Session {
-            id: SessionId::new(),
-            organization_id: "org_00000000000000000000000000000001".to_string(),
-            harness_id: HarnessId::new(),
-            agent_id: None,
-            title: Some(title.to_string()),
-            preview: None,
-            output_preview: None,
-            tags: vec![],
-            model_id: None,
-            capabilities: vec![],
-            tools: vec![],
-            status: SessionStatus::Idle,
-            created_at: Utc::now(),
-            updated_at: Utc::now(),
-            started_at: None,
-            finished_at: None,
-            usage: None,
-            is_pinned: None,
-            active_schedule_count: None,
-            features: vec![],
-        }
-    }
-
-    struct MockPlatformStore {
-        harness: Harness,
-        agent: Agent,
-        session: Session,
-    }
-
-    impl MockPlatformStore {
-        fn new() -> Self {
-            Self {
-                harness: build_harness("Test Harness"),
-                agent: build_agent("Test Agent"),
-                session: build_session("Test Session"),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl PlatformStore for MockPlatformStore {
-        async fn list_harnesses(&self) -> Result<Vec<Harness>> {
-            Ok(vec![self.harness.clone()])
-        }
-        async fn get_harness(&self, _id: HarnessId) -> Result<Option<Harness>> {
-            Ok(Some(self.harness.clone()))
-        }
-        async fn create_harness(
-            &self,
-            name: &str,
-            _desc: Option<&str>,
-            _prompt: &str,
-            _caps: &[String],
-        ) -> Result<Harness> {
-            let mut h = self.harness.clone();
-            h.name = name.to_string();
-            Ok(h)
-        }
-        async fn update_harness(
-            &self,
-            _id: HarnessId,
-            name: Option<&str>,
-            _desc: Option<&str>,
-            _prompt: Option<&str>,
-        ) -> Result<Harness> {
-            let mut h = self.harness.clone();
-            if let Some(n) = name {
-                h.name = n.to_string();
-            }
-            Ok(h)
-        }
-        async fn delete_harness(&self, _id: HarnessId) -> Result<()> {
-            Ok(())
-        }
-        async fn copy_harness(&self, _id: HarnessId, new_name: Option<&str>) -> Result<Harness> {
-            let mut h = self.harness.clone();
-            h.id = HarnessId::new();
-            h.name = new_name.unwrap_or("Copy").to_string();
-            Ok(h)
-        }
-        async fn list_agents(&self) -> Result<Vec<Agent>> {
-            Ok(vec![self.agent.clone()])
-        }
-        async fn get_agent_by_id(&self, _id: AgentId) -> Result<Option<Agent>> {
-            Ok(Some(self.agent.clone()))
-        }
-        async fn create_agent(
-            &self,
-            name: &str,
-            _desc: Option<&str>,
-            _prompt: &str,
-            _caps: &[String],
-        ) -> Result<Agent> {
-            let mut a = self.agent.clone();
-            a.name = name.to_string();
-            Ok(a)
-        }
-        async fn update_agent(
-            &self,
-            _id: AgentId,
-            name: Option<&str>,
-            _desc: Option<&str>,
-            _prompt: Option<&str>,
-        ) -> Result<Agent> {
-            let mut a = self.agent.clone();
-            if let Some(n) = name {
-                a.name = n.to_string();
-            }
-            Ok(a)
-        }
-        async fn delete_agent(&self, _id: AgentId) -> Result<()> {
-            Ok(())
-        }
-        async fn list_sessions(
-            &self,
-            _limit: Option<usize>,
-            _agent_id: Option<AgentId>,
-        ) -> Result<Vec<Session>> {
-            Ok(vec![self.session.clone()])
-        }
-        async fn create_session(
-            &self,
-            _hid: HarnessId,
-            _aid: Option<AgentId>,
-            title: Option<&str>,
-        ) -> Result<Session> {
-            let mut s = self.session.clone();
-            s.title = title.map(|t| t.to_string());
-            Ok(s)
-        }
-        async fn get_session_by_id(&self, _id: SessionId) -> Result<Option<Session>> {
-            Ok(Some(self.session.clone()))
-        }
-        async fn delete_session(&self, _id: SessionId) -> Result<()> {
-            Ok(())
-        }
-        async fn send_message(&self, _id: SessionId, _content: &str) -> Result<()> {
-            Ok(())
-        }
-        async fn get_messages(
-            &self,
-            _id: SessionId,
-            _limit: Option<usize>,
-        ) -> Result<Vec<PlatformMessage>> {
-            Ok(vec![
-                PlatformMessage {
-                    role: "user".into(),
-                    content: "Hello".into(),
-                    created_at: Utc::now(),
-                },
-                PlatformMessage {
-                    role: "agent".into(),
-                    content: "Hi!".into(),
-                    created_at: Utc::now(),
-                },
-            ])
-        }
-        async fn wait_for_idle(&self, _id: SessionId, _t: Option<u64>) -> Result<String> {
-            Ok("idle".to_string())
-        }
-        fn base_url(&self) -> &str {
-            "http://localhost:3000"
-        }
-    }
 
     fn mock_context() -> ToolContext {
         let store: Arc<dyn PlatformStore> = Arc::new(MockPlatformStore::new());
